requirements.txt
click==6.7
croniter==0.3.13
decorator==4.0.10
Django==1.10.5
django-bootstrap3==8.0.0
django-compressor==2.1.1
django-codemirror-widget==0.4.1
django-inline-actions==1.0.0
django-model-utils==2.6.1
django-modeltranslation==0.12
django-mptt==0.8.7
django-mptt-admin==0.4.4
django-picklefield==0.3.2
django-recaptcha2==1.0.0
django-rq==0.9.4
django-suit==0.2.23
django-suit-ckeditor==0.0.2
django-suit-rq==1.0.1
-e git://github.com/JP74/django-suit-sortable.git@d14f97a3480da7d0f1ca43b3d34568666952e03e#egg=django_suit_sortable
djangorestframework==3.5.4
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
olefile==0.44
pdfcrowd==2.7
pdfkit==0.6.1
pexpect==4.2.1
phonenumbers==8.0.0
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.6.0
pytils==0.3
pytz==2016.10
redis==2.10.5
requests==2.12.4
rq==0.7.1
rq-scheduler==0.7.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
Wand==0.4.4
wcwidth==0.1.7
xlrd==1.0.0
xlwt==1.2.0
git+https://github.com/arthanson/django-genericadmin.git
django-autocomplete-light==3.2.1

